refactoring : moved authMe request into session slice of entities layer. (#173)
