--- src/features/login/api/authApi.ts
@@ -1,25 +1,0 @@
-import { ApiSuccessResponse, baseAPI, ENDPOINTS, TAGS } from '@shared/api';
-
-interface AuthData {
-  userId: number;
-  username: string;
-  email: string;
-}
-
-type AuthMeResponse = ApiSuccessResponse<AuthData>;
-
-const authApi = baseAPI.injectEndpoints({
-  endpoints: (builder) => ({
-    getAuthData: builder.query<AuthMeResponse, void>({
-      query: () => ({
-        url: ENDPOINTS.AUTH_ME,
-        headers: {
-          // 'x-mock-response-code': '401',
-        },
-      }),
-      providesTags: [TAGS.AUTH],
-    }),
-  }),
-});
-
-export const { useGetAuthDataQuery } = authApi;
