+++ src/widgets/login-form/model/useLoginForm.ts
@@ -2,9 +2,12 @@ import { BaseSyntheticEvent, useEffect, useState } from 'react';
 import { SubmitHandler, useForm, UseFormReturn } from 'react-hook-form';
 import { ErrorMessage, getErrorMessage } from '@shared/error-message';
 import { useAppDispatch } from '@shared/redux';
-import { storeAccessToken } from '@entities/session';
-import { useLoginMutation } from '@entities/session/api/loginApi';
-import { LoginData, LoginField } from '@entities/session/model/types';
+import {
+  LoginData,
+  LoginField,
+  storeAccessToken,
+  useLoginMutation,
+} from '@entities/session';
 import { LOGIN_ERROR_MESSAGES } from '../consts/error-messages';
 
 interface UseLogInFormResult {
